add releasenote and docs config
This change stubs out the docs and release note
structure and set up the inintal jobs to build
and publish docs and releasenotes.

This invovle adding a bindep.txt to capture
the binary packages requried for html and pdf docs,
extending tox with the standard docs targets,
provideing a doc/requirement.txt to track the
doc only python deps and a standard folder structure.

Change-Id: Iffbef50ab806688a6d62f27ab60fa824edcb3d09