Add basic docs and runserver command
This commit provides some basic docs on how the
repo works. These shoudl be moved to contibutor
docs later once we have a devstack plugin and
a working install procedure. For now they
are stored in the README of discoverablity.

This change also provide a minimal settings.py
to be used in conjunction with a new
``tox -e runserver`` command. this will run
a local development instance of horizon using in memory
caching an signed cookies for session management.

in the future we may evolve this to run without
all of horizon or modify it based on how the plugin
develops once we have a working devstack jobs.

An example local.conf is also provided which deploys
nova, neutron, placement, keystone, glance an horizon.
addtionally promethus, ceilometer and sg-core are deployed
to provide metrics.

Change-Id: If9f38d88efccd29a63dd2aba4edd56f5f9fa1933